refactor: reorg ConfigManager.lua to Player/
